+++ src/views/Stake/Stake.tsx
@@ -1,204 +1,21 @@
 import "./Stake.scss";
 
-import { t, Trans } from "@lingui/macro";
-import { Box, Divider, Grid, Typography, Zoom } from "@material-ui/core";
-import { Skeleton } from "@material-ui/lab";
-import { InputWrapper, MetricCollection, Paper, PrimaryButton, Tab, Tabs } from "@olympusdao/component-library";
-import { ethers } from "ethers";
-import { ChangeEvent, ChangeEventHandler, memo, useCallback, useState } from "react";
-import { useDispatch } from "react-redux";
-import { useHistory } from "react-router";
-import ConnectButton from "src/components/ConnectButton/ConnectButton";
-import { useAppSelector } from "src/hooks";
-import { usePathForNetwork } from "src/hooks/usePathForNetwork";
-import { useWeb3Context } from "src/hooks/web3Context";
-import { isPendingTxn, txnButtonText } from "src/slices/PendingTxnsSlice";
+import { t } from "@lingui/macro";
+import { Grid, Zoom } from "@material-ui/core";
+import { MetricCollection, Paper } from "@olympusdao/component-library";
+import { memo, useState } from "react";
 
-import { getGohmBalFromSohm } from "../../helpers";
-import { error } from "../../slices/MessagesSlice";
-import { changeApproval, changeStake } from "../../slices/StakeThunk";
-import { changeApproval as changeGohmApproval } from "../../slices/WrapThunk";
 import { CurrentIndex, StakingAPY, TotalValueDeposited } from "../TreasuryDashboard/components/Metric/Metric";
 import ExternalStakePools from "./components/ExternalStakePools/ExternalStakePools";
 import RebaseTimer from "./components/RebaseTimer/RebaseTimer";
-import { StakeBalances } from "./components/StakeBalances";
-import { StakeFiveDayYield } from "./components/StakeFiveDayYield";
-import { StakeNextRebaseAmount } from "./components/StakeNextRebaseAmount";
-import { StakeRebaseYield } from "./components/StakeRebaseYield";
-import { ConfirmDialog } from "./ConfirmDialog";
+import { StakeActionArea } from "./components/StakeActionArea/StakeActionArea";
 
 const Stake: React.FC = () => {
-  const dispatch = useDispatch();
-  const history = useHistory();
-  const { provider, address, networkId } = useWeb3Context();
-  usePathForNetwork({ pathName: "stake", networkID: networkId, history });
-
-  const [zoomed, setZoomed] = useState(false);
-  const [view, setView] = useState(0);
-  const [quantity, setQuantity] = useState("");
-  const [confirmation, setConfirmation] = useState(false);
-
-  const currentIndex = useAppSelector(state => {
-    return state.app.currentIndex;
-  });
-
-  const ohmBalance = useAppSelector(state => {
-    return state.account.balances && state.account.balances.ohm;
-  });
-  const sohmBalance = useAppSelector(state => {
-    return state.account.balances && state.account.balances.sohm;
-  });
-
-  const gOhmAsSohm = useAppSelector(state => {
-    return state.account.balances && state.account.balances.gOhmAsSohmBal;
-  });
-
-  const stakeAllowance = useAppSelector(state => {
-    return (state.account.staking && state.account.staking.ohmStake) || 0;
-  });
-  const unstakeAllowance = useAppSelector(state => {
-    return (state.account.staking && state.account.staking.ohmUnstake) || 0;
-  });
-
-  const directUnstakeAllowance = useAppSelector(state => {
-    return (state.account.wrapping && state.account.wrapping.gOhmUnwrap) || 0;
-  });
-
-  const pendingTransactions = useAppSelector(state => {
-    return state.pendingTransactions;
-  });
-
-  const setMax = () => {
-    if (view === 0) {
-      setQuantity(ohmBalance);
-    } else if (!confirmation) {
-      setQuantity(sohmBalance);
-    } else if (confirmation) {
-      setQuantity(gOhmAsSohm.toString());
-    }
-  };
-
-  const onSeekApproval = async (token: string) => {
-    if (token === "gohm") {
-      await dispatch(changeGohmApproval({ address, token: token.toLowerCase(), provider, networkID: networkId }));
-    } else {
-      await dispatch(changeApproval({ address, token, provider, networkID: networkId, version2: true }));
-    }
-  };
-
-  const onChangeStake = async (action: string) => {
-    // eslint-disable-next-line no-restricted-globals
-    if (isNaN(Number(quantity)) || Number(quantity) === 0 || Number(quantity) < 0) {
-      // eslint-disable-next-line no-alert
-      return dispatch(error(t`Please enter a value!`));
-    }
-
-    // 1st catch if quantity > balance
-    const gweiValue = ethers.utils.parseUnits(quantity.toString(), "gwei");
-    if (action === "stake" && gweiValue.gt(ethers.utils.parseUnits(ohmBalance, "gwei"))) {
-      return dispatch(error(t`You cannot stake more than your OHM balance.`));
-    }
-
-    if (confirmation === false && action === "unstake" && gweiValue.gt(ethers.utils.parseUnits(sohmBalance, "gwei"))) {
-      return dispatch(
-        error(
-          t`You do not have enough sOHM to complete this transaction.  To unstake from gOHM, please toggle the sohm-gohm switch.`,
-        ),
-      );
-    }
-
-    /**
-     * converts sOHM quantity to gOHM quantity when box is checked for gOHM staking
-     * @returns sOHM as gOHM quantity
-     */
-    // const formQuant = checked && currentIndex && view === 1 ? quantity / Number(currentIndex) : quantity;
-    const formQuant = async () => {
-      if (confirmation && currentIndex && view === 1) {
-        return await getGohmBalFromSohm({ provider, networkID: networkId, sOHMbalance: quantity });
-      } else {
-        return quantity;
-      }
-    };
-
-    await dispatch(
-      changeStake({
-        address,
-        action,
-        value: await formQuant(),
-        provider,
-        networkID: networkId,
-        version2: true,
-        rebase: !confirmation,
-      }),
-    );
-  };
-
-  const hasAllowance = useCallback(
-    token => {
-      if (token === "ohm") return stakeAllowance > 0;
-      if (token === "sohm") return unstakeAllowance > 0;
-      if (token === "gohm") return directUnstakeAllowance > 0;
-      return 0;
-    },
-    [stakeAllowance, unstakeAllowance, directUnstakeAllowance],
-  );
-
-  const isAllowanceDataLoading = (stakeAllowance == null && view === 0) || (unstakeAllowance == null && view === 1);
-
-  const changeView: any = (_event: ChangeEvent<any>, newView: number) => {
-    setView(newView);
-  };
-
-  const handleChangeQuantity = useCallback<ChangeEventHandler<HTMLInputElement>>(e => {
-    if (Number(e.target.value) >= 0) setQuantity(e.target.value);
-  }, []);
-
-  let stakeOnClick: () => Promise<{ payload: string; type: string } | undefined | void>;
-  let stakeDisabled: boolean;
-  let stakeButtonText: string;
-
-  //set defaults. if unstake tab selected else use staking tab as default
-  if (view === 1) {
-    stakeDisabled = isPendingTxn(pendingTransactions, "approve_unstaking");
-    stakeOnClick = () => onSeekApproval(confirmation ? "gohm" : "sohm");
-    stakeButtonText = txnButtonText(pendingTransactions, "approve_unstaking", t`Approve`);
-  } else {
-    stakeDisabled = isPendingTxn(pendingTransactions, "approve_staking");
-    stakeOnClick = () => onSeekApproval("ohm");
-    stakeButtonText = txnButtonText(pendingTransactions, "approve_staking", t`Approve`);
-  }
-
-  //evaluate if data allowance data is finished loading
-  if (!isAllowanceDataLoading) {
-    //If Staking Tab
-    if (view === 0) {
-      if (address && hasAllowance("ohm")) {
-        stakeDisabled = isPendingTxn(pendingTransactions, "staking");
-        stakeOnClick = () => onChangeStake("stake");
-        stakeButtonText = txnButtonText(
-          pendingTransactions,
-          "staking",
-          `${t`Stake to`} ${confirmation ? " gOHM" : " sOHM"}`,
-        );
-      }
-    }
-    //If Unstaking Tab
-    if (view === 1) {
-      if ((address && hasAllowance("sohm") && !confirmation) || (hasAllowance("gohm") && confirmation)) {
-        stakeDisabled = isPendingTxn(pendingTransactions, "unstaking");
-        stakeOnClick = () => onChangeStake("unstake");
-        stakeButtonText = txnButtonText(
-          pendingTransactions,
-          "unstaking",
-          `${t`Unstake from`} ${confirmation ? " gOHM" : " sOHM"}`,
-        );
-      }
-    }
-  }
+  const [isZoombed, setIsZoomed] = useState(false);
 
   return (
     <div id="stake-view">
-      <Zoom in={true} onEntered={() => setZoomed(true)}>
+      <Zoom in onEntered={() => setIsZoomed(true)}>
         <Paper headerText={t`Single Stake (3, 3)`} subHeader={<RebaseTimer />}>
           <Grid container direction="column" spacing={2}>
             <Grid item>
@@ -210,120 +27,8 @@ const Stake: React.FC = () => {
                 <CurrentIndex className="stake-index" />
               </MetricCollection>
             </Grid>
-            <div className="staking-area">
-              {!address ? (
-                <div className="stake-wallet-notification">
-                  <div className="wallet-menu" id="wallet-menu">
-                    <ConnectButton />
-                  </div>
-                  <Typography variant="h6">
-                    <Trans>Connect your wallet to stake OHM</Trans>
-                  </Typography>
-                </div>
-              ) : (
-                <>
-                  <Box className="stake-action-area">
-                    <Tabs
-                      key={String(zoomed)}
-                      centered
-                      value={view}
-                      textColor="primary"
-                      indicatorColor="primary"
-                      className="stake-tab-buttons"
-                      onChange={changeView}
-                      aria-label="stake tabs"
-                      //hides the tab underline sliding animation in while <Zoom> is loading
-                      TabIndicatorProps={!zoomed ? { style: { display: "none" } } : undefined}
-                    >
-                      <Tab
-                        aria-label="stake-button"
-                        label={t({
-                          id: "do_stake",
-                          comment: "The action of staking (verb)",
-                        })}
-                      />
-                      <Tab aria-label="unstake-button" label={t`Unstake`} />
-                    </Tabs>
-                    <Grid container className="stake-action-row">
-                      {address && !isAllowanceDataLoading ? (
-                        (!hasAllowance("ohm") && view === 0) ||
-                        (!hasAllowance("sohm") && view === 1 && !confirmation) ||
-                        (!hasAllowance("gohm") && view === 1 && confirmation) ? (
-                          <>
-                            <Grid item xs={12} sm={8} className="stake-grid-item">
-                              <Box mt={"10px"}>
-                                <Typography variant="body1" className="stake-note" color="textSecondary">
-                                  {view === 0 ? (
-                                    <>
-                                      <Trans>First time staking</Trans> <b>OHM</b>?
-                                      <br />
-                                      <Trans>Please approve Olympus Dao to use your</Trans> <b>OHM</b>{" "}
-                                      <Trans>for staking</Trans>.
-                                    </>
-                                  ) : (
-                                    <>
-                                      <Trans>First time unstaking</Trans> <b>{confirmation ? "gOHM" : "sOHM"}</b>?
-                                      <br />
-                                      <Trans>Please approve Olympus Dao to use your</Trans>{" "}
-                                      <b>{confirmation ? "gOHM" : "sOHM"}</b> <Trans>for unstaking</Trans>.
-                                    </>
-                                  )}
-                                </Typography>
-                              </Box>
-                            </Grid>
-                            <Grid item xs={12} sm={4} className="stake-grid-item">
-                              <Box mt={1}>
-                                <PrimaryButton
-                                  fullWidth
-                                  className="stake-button"
-                                  disabled={stakeDisabled}
-                                  onClick={stakeOnClick}
-                                >
-                                  {stakeButtonText}
-                                </PrimaryButton>
-                              </Box>
-                            </Grid>
-                          </>
-                        ) : (
-                          <InputWrapper
-                            id="amount-input"
-                            type="number"
-                            label={t`Enter an amount`}
-                            value={quantity}
-                            onChange={handleChangeQuantity}
-                            labelWidth={0}
-                            endString={t`Max`}
-                            endStringOnClick={setMax}
-                            buttonText={stakeButtonText}
-                            buttonOnClick={stakeOnClick}
-                            disabled={stakeDisabled}
-                          />
-                        )
-                      ) : (
-                        <Skeleton width="150px" />
-                      )}
-                    </Grid>
-                  </Box>
-                  <ConfirmDialog
-                    quantity={quantity}
-                    currentIndex={currentIndex}
-                    view={view}
-                    onConfirm={setConfirmation}
-                  />
-                  <div className="stake-user-data">
-                    <StakeBalances />
 
-                    <Divider color="secondary" />
-
-                    <StakeNextRebaseAmount />
-
-                    <StakeRebaseYield />
-
-                    <StakeFiveDayYield />
-                  </div>
-                </>
-              )}
-            </div>
+            <StakeActionArea isZoomed={isZoombed} />
           </Grid>
         </Paper>
       </Zoom>
