+++ src/views/Stake/components/ConfirmDialog/__tests__/ConfirmDialog.unit.test.tsx
@@ -1,4 +1,5 @@
-import { render } from "../../../testUtils";
+import { render } from "src/testUtils";
+
 import { ConfirmDialog } from "../ConfirmDialog";
 
 describe("<ConfirmDialog/>", () => {
